refactor: rename logs to web3-logs
