+++ .gitignore
@@ -33,6 +33,7 @@ compilation-stats.json
 /contracts
 
 # Following content is copied from https://github.com/github/gitignore/blob/master/Node.gitignore
+logs
 *.log
 npm-debug.log*
 yarn-debug.log*
