+++ .prettierignore
@@ -42,6 +42,7 @@ dist/*
 
 # Following content is copied from https://github.com/github/gitignore/blob/master/Node.gitignore
 # Logs
+logs
 *.log
 npm-debug.log*
 yarn-debug.log*
