+++ packages/dashboard/package.json
@@ -31,7 +31,7 @@
     "@masknet/shared-base-ui": "workspace:^",
     "@masknet/storybook-shared": "workspace:^",
     "@masknet/theme": "workspace:^",
-    "@masknet/logs": "workspace:^",
+    "@masknet/web3-logs": "workspace:^",
     "@masknet/web3-helpers": "workspace:^",
     "@masknet/web3-hooks-base": "workspace:^",
     "@masknet/web3-hooks-evm": "workspace:^",
