+++ packages/dashboard/src/initialization/Dashboard.tsx
@@ -10,9 +10,9 @@ import {
     DialogStackingProvider,
 } from '@masknet/theme'
 import { I18NextProviderHMR, SharedContextProvider } from '@masknet/shared'
-import { LogContextProvider } from '@masknet/logs/hooks'
 import { ErrorBoundary } from '@masknet/shared-base-ui'
 import { createInjectHooksRenderer, useActivatedPluginsDashboard } from '@masknet/plugin-infra/dashboard'
+import { LogContextProvider } from '@masknet/web3-logs/hooks'
 import { EnvironmentContextProvider, Web3ContextProvider } from '@masknet/web3-hooks-base'
 import { i18NextInstance, NetworkPluginID, queryRemoteI18NBundle } from '@masknet/shared-base'
