+++ packages/dashboard/tsconfig.json
@@ -10,10 +10,10 @@
   "references": [
     { "path": "../public-api" },
     { "path": "../theme/" },
-    { "path": "../logs/" },
     { "path": "../mask/background" },
     { "path": "../web3-hooks/base/" },
     { "path": "../web3-hooks/evm/" },
+    { "path": "../web3-logs" },
     { "path": "../web3-shared/base/" },
     { "path": "../web3-shared/evm/" },
     { "path": "../web3-shared/flow/" },
