+++ packages/mask/package.json
@@ -23,7 +23,6 @@
     "@masknet/gun-utils": "workspace:^",
     "@masknet/icons": "workspace:^",
     "@masknet/injected-script": "workspace:^",
-    "@masknet/logs": "workspace:^",
     "@masknet/plugin-approval": "workspace:^",
     "@masknet/plugin-cross-chain-bridge": "workspace:^",
     "@masknet/plugin-cyberconnect": "workspace:^",
@@ -64,6 +63,7 @@
     "@masknet/web3-helpers": "workspace:^",
     "@masknet/web3-hooks-base": "workspace:^",
     "@masknet/web3-hooks-evm": "workspace:^",
+    "@masknet/web3-logs": "workspace:^",
     "@masknet/web3-providers": "workspace:^",
     "@masknet/web3-shared-base": "workspace:^",
     "@masknet/web3-shared-evm": "workspace:^",
