+++ packages/mask/src/extension/popups/UI.tsx
@@ -2,11 +2,11 @@ import { lazy, useEffect, useState, useMemo } from 'react'
 import { Navigate, Route, Routes, HashRouter } from 'react-router-dom'
 import { createInjectHooksRenderer, useActivatedPluginsDashboard } from '@masknet/plugin-infra/dashboard'
 import { NetworkPluginID, PopupRoutes, queryRemoteI18NBundle } from '@masknet/shared-base'
-import { LogContextProvider } from '@masknet/logs/hooks'
 import { useValueRef } from '@masknet/shared-base-ui'
 import { PopupSnackbarProvider } from '@masknet/theme'
 import { Web3ContextProvider } from '@masknet/web3-hooks-base'
 import { ProviderType } from '@masknet/web3-shared-evm'
+import { LogContextProvider } from '@masknet/web3-logs/hooks'
 import { usePopupFullPageTheme } from '../../utils/theme/useClassicMaskFullPageTheme.js'
 import { languageSettings } from '../../../shared/legacy-settings/settings.js'
 import { LoadingPlaceholder } from './components/LoadingPlaceholder/index.js'
