+++ packages/plugins/Debugger/package.json
@@ -12,7 +12,6 @@
     "@masknet/icons": "workspace:^",
     "@masknet/plugin-infra": "workspace:^",
     "@masknet/plugin-wallet": "workspace:^",
-    "@masknet/logs": "workspace:^",
     "@masknet/shared": "workspace:^",
     "@masknet/shared-base": "workspace:^",
     "@masknet/shared-base-ui": "workspace:^",
@@ -21,6 +20,7 @@
     "@masknet/web3-contracts": "workspace:^",
     "@masknet/web3-hooks-base": "workspace:^",
     "@masknet/web3-hooks-evm": "workspace:^",
+    "@masknet/web3-logs": "workspace:^",
     "@masknet/web3-providers": "workspace:^",
     "@masknet/web3-shared-base": "workspace:^",
     "@masknet/web3-shared-evm": "workspace:^",
