+++ packages/plugins/Debugger/tsconfig.json
@@ -12,7 +12,7 @@
     { "path": "../../web3-helpers" },
     { "path": "../../web3-hooks/base" },
     { "path": "../../web3-hooks/evm" },
-    { "path": "../../logs" },
+    { "path": "../../web3-logs" },
     { "path": "../../shared" },
     { "path": "../../shared-base-ui" }
   ]
