+++ packages/web3-logs/tsconfig.json
