+++ tsconfig.json
@@ -79,7 +79,6 @@
       "@masknet/global-types/*": ["./packages/polyfills/types/*"],
       "@masknet/public-api": ["./packages/public-api/src"],
       // @masknet/scripts is not listed. It does not mean to be installed as a dependency and not join the type check.
-      "@masknet/logs": ["./packages/logs/src"],
       "@masknet/shared": ["./packages/shared/src"],
       "@masknet/shared-base": ["./packages/shared-base/src"],
       "@masknet/shared-base-ui": ["./packages/shared-base-ui/src"],
@@ -87,6 +86,7 @@
       "@masknet/theme": ["./packages/theme/src"],
       // @masknet/web3-constants is not listed. It is not a TS project (only contains JSON files).
       // @masknet/web3-contracts is not listed. It is not a TS project (only contains generated JSON and .d.ts files).
+      "@masknet/web3-logs": ["packages/web3-logs/src"],
       "@masknet/web3-helpers": ["./packages/web3-helpers/src"],
       "@masknet/web3-providers": ["./packages/web3-providers/src"],
       "@masknet/web3-state": ["./packages/web3-state/src"],
